Slit tests for NexusLinkMiddleware into the separate features.
